% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_sis.R
\name{abc_sis}
\alias{abc_sis}
\title{ABC sampler for the SIS model}
\usage{
abc_sis(
  A_mat,
  prevalence = 0.025,
  days_infectious = 7,
  beta_scalar_range = c(1.25, 1.5),
  n_draws = 5000,
  sis_nsims = 300,
  showProgress = TRUE,
  cl = NULL
)
}
\arguments{
\item{A_mat}{nxn adjacency matrix}

\item{prevalence}{scalar.}

\item{days_infectious}{integer.}

\item{beta_scalar_range}{vector with lower and upper bounds for the uniform prior 
on the transmission parameter beta}

\item{n_draws}{How many draws to obtain}

\item{sis_nsims}{How long should the SIS sampler run (in days)?}

\item{showProgress}{logical.  Should a progress bar be displayed?}

\item{cl}{If running in parallel using the parallel package, pass 
in object from parallel::makeCluster()}
}
\value{
abc_sis returns a n_draws x (1 + n) matrix.  The first 
column is the draws of beta, the remaining columns correspond 
to \eqn{y_1,\ldots,y_n}.
}
\description{
This function runs an approximate Bayesian
computation (ABC) algorithm to obtain samples
from the SIS model, yielding samples of y
given prevalence within a small window of
that specified.
}
