% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pools.R
\name{get_pools}
\alias{get_pools}
\title{Obtain optimal pooling assignments}
\usage{
get_pools(
  A,
  sensitivity = 0.75,
  specificity = 0.995,
  K = 10,
  y_draws,
  Z_initial,
  iter_per_temp = 100,
  maxIterConstant = 1000,
  showProgress = TRUE,
  temperatures = 2 * 0.95^c(1:500 - 1),
  extra_individuals = c("combine", "new_cluster")[1]
)
}
\arguments{
\item{A}{igraph object.  Network conveying transmission opportunities between
individuals}

\item{sensitivity}{numeric.}

\item{specificity}{numeric.}

\item{K}{integer. Number of individuals in each pool}

\item{y_draws}{matrix. (# draws) x (# individuals) matrix.  Sparse format
(dgCMatrix) is highly recommended for computational efficiency.}

\item{Z_initial}{(optional) matrix giving pooling assignments to initialize
the simulated annealing algorithm. If left blank (recommended), a constrained 
PAM clustering algorithm will be used.}

\item{iter_per_temp}{integer. Number of iterations per SA temperature}

\item{maxIterConstant}{integer.  Number of iterations with no change before 
stopping the SA algorithm.}

\item{showProgress}{logical.  Should a progress bar be displayed?}

\item{temperatures}{positive monotonically decreasing numeric vector.  SA 
cooling schedule}

\item{extra_individuals}{character describing what to do with the remainder of
n/P. Either "combine" to evenly distribute the remainder across existing pools, 
or "new_cluster" to create a new pool of less than K individuals.}
}
\value{
get_pools returns a list with named elements Z, the nxP matrix giving
the optimal pool assignments, and fvalues, the vector of the objective function
over the SA iterations.
}
\description{
This functions takes as an input a network and a matrix of
data generated from some transmission model (e.g., SIS model),
and outputs the optimal pooling assignment as determined 
through a simulated annealing algorithm.
}
