% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malinovsky.R
\name{malinovsky}
\alias{malinovsky}
\title{Obtain E(C)/E(T) without network information}
\usage{
malinovsky(q, Sp, Se, k)
}
\arguments{
\item{q}{numeric between 0 and 1, giving 1 minus the prevalence.}

\item{Sp}{numeric between 0 and 1, giving the specificity}

\item{Se}{numeric between 0 and 1, giving the sensitivity}

\item{k}{integer The number of individuals in each pool}
}
\value{
named list with value, the objective function, 
the numerator, the expected number of correctly classified 
individuals, and the denominator, the expected number
of tests.
}
\description{
malinovsky computes the objective function in 
Malinovsky et al. (2015) <https://doi.org/10.1111/biom.12385>
This gives the expected number of correctly classified
individuals divided by the expected number of tests.
}
