\name{networkPooledTesting-package}
\alias{networkPooledTesting-package}
\alias{networkPooledTesting}
\docType{package}
\title{
\packageTitle{networkPooledTesting}
}
\description{
\packageDescription{networkPooledTesting}
}
\details{

\packageDESCRIPTION{networkPooledTesting}
\packageIndices{networkPooledTesting}
The primary function of this package is get_pools().  This requires a network in igraph format, as well as a matrix of data generated according to some transmission model (or other applicable data generating mechanism if not working in the infectious disease setting).  The sis_simulator(), and its nicer wrapper function abc_sis(), can be used for this purpose, although really nearly anything is acceptable.
}
\author{
\packageAuthor{networkPooledTesting}

\packageMaintainer{networkPooledTesting}
}
\references{
Sewell, Daniel K. (2021+). "Leveraging network structure to improve pooled testing efficiency," Journal of the Royal Statistical Society: Series C (Applied Statistics).

Malinovsky, Y., Albert, P.S. and Roy, A. (2016), Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification. Biom, 72: 299-302.
}
