% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sis_simulator}
\alias{sis_simulator}
\title{SIS Simulator}
\usage{
sis_simulator(max_time, initial, A_mat, beta, days_infectious)
}
\arguments{
\item{max_time}{Integer. Number of days to simulate.}

\item{initial}{1xn binary vector indicating initially infected individuals.}

\item{A_mat}{Sparse adjacency matrix.}

\item{beta}{positive double. Transmission parameter.}

\item{days_infectious}{integer.  Number of days an individual remains infectious.}
}
\value{

}
\description{
This function simulates data according to the 
susceptible-infected-susceptible (SIS) 
compartmental model.
}
